# shared/openapi/openapi.yaml
# Canonical public API spec — OpenAI-compatible surface.
# Generated TypeScript SDK lives at frontend-web/src/shared/api/generated/.
# Generated Python types live at backend-python/app/api/generated/.
# Single source of truth: this file. Both languages run codegen on CI.

openapi: 3.1.0
info:
  title: OpenRouter Gateway API
  version: 1.0.0
  description: |
    OpenAI-compatible unified gateway. Drop-in replacement: change only `base_url`.
    See DEVELOPMENT_PLAN.md for module mapping.
  contact:
    name: Engineering
    url: https://openrouter.example.com/contact

servers:
  - url: https://api.openrouter.example.com/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local dev

security:
  - bearerAuth: []

paths:
  /chat/completions:
    post:
      operationId: createChatCompletion
      summary: Create a chat completion (DRD GW-001)
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Completion response (or SSE stream when stream=true)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
            text/event-stream:
              schema:
                type: string
                description: SSE stream of ChatCompletionChunk objects
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '402': { $ref: '#/components/responses/InsufficientCredits' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ProvidersUnavailable' }

  /models:
    get:
      operationId: listModels
      summary: List available models with pricing (DRD BL-011, MK-010)
      tags: [models]
      responses:
        '200':
          description: Catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Model' }

  /embeddings:
    post:
      operationId: createEmbedding
      summary: Create embeddings (DRD §8.5)
      tags: [embeddings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: Embedding response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbeddingResponse' }

  /images/generations:
    post:
      operationId: createImage
      summary: Generate image (DRD §8.2)
      tags: [images]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImageGenerationRequest' }
      responses:
        '200':
          description: Image response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImageGenerationResponse' }

  /videos:
    post:
      operationId: submitVideoJob
      summary: Submit async video generation job (DRD §8.3)
      tags: [videos]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VideoJobRequest' }
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideoJob' }

  /videos/{job_id}:
    get:
      operationId: getVideoJob
      summary: Poll video job status
      tags: [videos]
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideoJob' }

  /audio/transcriptions:
    post:
      operationId: createTranscription
      summary: Speech-to-text (DRD §8.4)
      tags: [audio]
      responses:
        '200':
          description: Transcription
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TranscriptionResponse' }

  /rerank:
    post:
      operationId: rerankDocuments
      summary: Rerank documents (DRD §8.6)
      tags: [reranking]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RerankRequest' }
      responses:
        '200':
          description: Reranked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RerankResponse' }

  /health:
    get:
      operationId: healthCheck
      summary: Service + dependencies health
      tags: [system]
      security: []
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HealthResponse' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key — sk-or-v1-...

  schemas:
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model: { type: string, example: "anthropic/claude-sonnet-4-20250514" }
        messages:
          type: array
          items: { $ref: '#/components/schemas/ChatMessage' }
        stream: { type: boolean, default: false }
        temperature: { type: number, minimum: 0, maximum: 2 }
        top_p: { type: number, minimum: 0, maximum: 1 }
        max_tokens: { type: integer, minimum: 1 }
        tools:
          type: array
          items: { $ref: '#/components/schemas/Tool' }
        tool_choice:
          oneOf:
            - { type: string, enum: [auto, none, required] }
            - { type: object }
        response_format: { $ref: '#/components/schemas/ResponseFormat' }
        reasoning_effort: { type: string, enum: [low, medium, high] }
        provider:
          type: object
          description: Provider-specific passthrough overrides (DRD PI-003)
          additionalProperties: true

    ChatMessage:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [system, user, assistant, tool] }
        content:
          oneOf:
            - { type: string }
            - type: array
              items: { $ref: '#/components/schemas/ContentBlock' }
        name: { type: string }
        tool_calls:
          type: array
          items: { $ref: '#/components/schemas/ToolCall' }
        tool_call_id: { type: string }

    ContentBlock:
      type: object
      required: [type]
      properties:
        type: { type: string, enum: [text, image_url, audio, video, document] }
        text: { type: string }
        image_url:
          type: object
          properties:
            url: { type: string }
            detail: { type: string, enum: [low, high, auto] }

    Tool:
      type: object
      required: [type, function]
      properties:
        type: { type: string, enum: [function] }
        function:
          type: object
          required: [name]
          properties:
            name: { type: string }
            description: { type: string }
            parameters: { type: object }

    ToolCall:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [function] }
        function:
          type: object
          properties:
            name: { type: string }
            arguments: { type: string }

    ResponseFormat:
      type: object
      properties:
        type: { type: string, enum: [text, json_object, json_schema] }
        json_schema: { type: object }

    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices, usage]
      properties:
        id: { type: string }
        object: { type: string, enum: [chat.completion] }
        created: { type: integer }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message: { $ref: '#/components/schemas/ChatMessage' }
              finish_reason: { type: string, enum: [stop, length, tool_calls, content_filter, error] }
        usage: { $ref: '#/components/schemas/Usage' }

    Usage:
      type: object
      properties:
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }
        cached_tokens: { type: integer }
        reasoning_tokens: { type: integer }

    Model:
      type: object
      required: [id, provider, modalities, pricing]
      properties:
        id: { type: string }
        provider: { type: string }
        modalities: { type: array, items: { type: string } }
        context_window: { type: integer }
        pricing:
          type: object
          properties:
            input_per_million: { type: number }
            output_per_million: { type: number }
            cached_input_per_million: { type: number }

    EmbeddingRequest:
      type: object
      required: [model, input]
      properties:
        model: { type: string }
        input:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
        dimensions: { type: integer }

    EmbeddingResponse:
      type: object
      properties:
        object: { type: string, enum: [list] }
        data:
          type: array
          items:
            type: object
            properties:
              object: { type: string, enum: [embedding] }
              embedding: { type: array, items: { type: number } }
              index: { type: integer }
        model: { type: string }
        usage: { $ref: '#/components/schemas/Usage' }

    ImageGenerationRequest:
      type: object
      required: [model, prompt]
      properties:
        model: { type: string }
        prompt: { type: string }
        n: { type: integer, default: 1 }
        size: { type: string }
        quality: { type: string, enum: [standard, hd] }
        response_format: { type: string, enum: [url, b64_json], default: url }

    ImageGenerationResponse:
      type: object
      properties:
        created: { type: integer }
        data:
          type: array
          items:
            type: object
            properties:
              url: { type: string }
              b64_json: { type: string }

    VideoJobRequest:
      type: object
      required: [model, prompt]
      properties:
        model: { type: string }
        prompt: { type: string }
        duration_seconds: { type: number }
        resolution: { type: string }
        aspect_ratio: { type: string }
        fps: { type: integer }

    VideoJob:
      type: object
      properties:
        job_id: { type: string }
        status: { type: string, enum: [pending, processing, completed, failed] }
        model: { type: string }
        created_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        unsigned_urls: { type: array, items: { type: string } }
        cost_usd: { type: number }
        error: { $ref: '#/components/schemas/Error' }

    TranscriptionResponse:
      type: object
      properties:
        text: { type: string }
        language: { type: string }
        duration: { type: number }

    RerankRequest:
      type: object
      required: [model, query, documents]
      properties:
        model: { type: string }
        query: { type: string }
        documents:
          type: array
          items: { type: string }
        top_n: { type: integer }

    RerankResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              relevance_score: { type: number }
              document: { type: string }

    HealthResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [healthy, degraded, unhealthy] }
        version: { type: string }
        dependencies:
          type: object
          additionalProperties:
            type: object
            properties:
              status: { type: string }
              latency_ms: { type: integer }

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type: { type: string }
            code: { type: string }
            param: { type: string }

  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Invalid or revoked API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InsufficientCredits:
      description: Account out of credits or budget exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
        X-RateLimit-Limit: { schema: { type: integer } }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ProvidersUnavailable:
      description: All providers in fallback chain failed
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }